+++ src/main/java/uk/org/openbanking/datamodel/service/converter/account/OBExternalAccountIdentificationConverter.java
@@ -18,7 +18,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package uk.org.openbanking.datamodel.service.converter;
+package uk.org.openbanking.datamodel.service.converter.account;
 
 import uk.org.openbanking.datamodel.account.OBExternalAccountIdentification3Code;
 import uk.org.openbanking.datamodel.account.OBExternalAccountIdentification4Code;
