+++ src/test/java/uk/org/openbanking/datamodel/service/converter/account/OBExternalAccountIdentificationConverterTest.java
@@ -18,7 +18,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package uk.org.openbanking.datamodel.service.converter;
+package uk.org.openbanking.datamodel.service.converter.account;
 
 import org.junit.Test;
 import uk.org.openbanking.datamodel.account.OBExternalAccountIdentification3Code;
